Develop

+++ app/Http/Controllers/HomeController.php
@@ -33,9 +33,9 @@ class HomeController extends Controller
         $direction = OrderByAdapter::toDirection($request);
         $internalOrderByKey = OrderByAdapter::toInternalKey($orderBy);
 
-        $limit = config('services.post.pagination.limit');
+        $limit = (int) config('services.post.pagination.limit');
         $page = PaginationRequestAdapter::getPage($request);
-        $offset = ($page * $limit) - $limit;
+        $offset = (int) (($page * $limit) - $limit);
         $itemsCount = $this->postManager->countForUser($currentUser->id);
         $totalPages = 1;
